Disable read data sources after alter rule and data source
